new, MangaDex Manga Reader

+++ commands/3.Moderasi/unelm.js
@@ -1,4 +1,5 @@
 const Discord = require('discord.js');
+const db = require('quick.db');
 
 exports.run = async (client, message, args) => {
   try {
@@ -8,17 +9,24 @@ exports.run = async (client, message, args) => {
     let korban = message.mentions.members.first() || message.guild.members.cache.get(args[0]);
     if (!korban) return message.channel.send("tag user yang ingin di unelm!");
 
+    //cari data
+    let data = new db.table('ELMs')
+    let elm = await data.fetch(korban.user.id)
+    if (elm === null) return;
+    let elm_role = data.get(korban.user.id)
 
-    let psantai = message.guild.roles.cache.get("647778783314837507");
-    korban.roles.add(psantai)
+    for (let i = 0; i < elm_role.length; i++) {
+      korban.roles.add(elm_role[i])
+    }
 
     let ELM = message.guild.roles.cache.get("505004825621168128");
-
     korban.roles.remove(ELM).then(() => {
       message.delete()
 
       message.channel.send(`**${korban.user.username}#${korban.user.discriminator}** telah selesai di unelm.`)
     })
+
+    await data.delete(korban.user.id)
 
     let embed = new Discord.MessageEmbed()
       .setAuthor(`UNELM | ${korban.user.tag}`)
